feat(website): clear up the layout mess
