--- src/app/(website)/loading.tsx
@@ -1,5 +1,0 @@
-import { Loading } from "@khenzii-dev/ui/atoms";
-
-const Skeleton = () => <Loading size={200} />;
-
-export default Skeleton;
