+++ src/app/not-found.tsx
@@ -1,0 +1,6 @@
+const NotFound = () => (
+    <p>Hello!</p>
+);
+
+export default NotFound;
+
