+++ src/styles/layout.module.scss
@@ -1,0 +1,8 @@
+@import "./constants.scss";
+
+.layout {
+    background-color: $background;
+    overflow-x: hidden;
+    margin: 0;
+}
+
